requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
fastapi==0.136.0
h11==0.16.0
idna==3.11
kombu==5.6.2
packaging==26.1
prompt_toolkit==3.0.52
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
pymongo==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
redis==7.4.0
requests==2.33.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.45.0
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
wsproto==1.3.2
